src/test/requirements.txt
pytest>=8
PyYAML>=6.0
psutil>=5.9
fabric>=3


